migrate unit tests for ppms entity
